pyproject.toml
[project]
name = "conda-portable"
version = "0.1.0"
description = "Make Conda environment.yml files portable across platforms"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
dependencies = [
  "PyYAML>=6.0",
  "conda-lock>=2.0",
]

authors = [
  { name = "Shekhar Narayanan", email = "s.narayanan@tilburguniversity.edu" }
]

[project.urls]
Homepage = "https://github.com/ShekharNarayanan/conda-portable"
Issues = "https://github.com/ShekharNarayanan/conda-portable/issues"
Source = "https://github.com/ShekharNarayanan/conda-portable"

[project.scripts]
conda-portable = "conda_portable.cli:main"  # CLI entry point

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.package-data]
conda_portable = ["common_packages.yaml"]
